backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Vector database and embeddings
pinecone-client==2.2.4
sentence-transformers==2.2.2
openai==1.3.8
tiktoken==0.5.2

# Document processing
langchain==0.0.350
langchain-community==0.0.13
unstructured==0.11.8
pypdf==3.17.4
beautifulsoup4==4.12.2
requests==2.31.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
ruff==0.1.6
black==23.11.0